fix(desktop): connect PwrGit through shared OAuth bridge
